Вынес константы, исправил недочеты
