+++ js/constants.js
@@ -1,0 +1,31 @@
+const MAX_COUNT_PHOTOS = 25;
+const MAX_COUNT_COMMENTS = 8;
+
+const COUNT_LIKES = {
+  MIN_LIKES : 15,
+  MAX_LIKES : 200,
+};
+
+const AVATAR_ID = {
+  MIN_ID : 1,
+  MAX_ID : 6,
+};
+
+const DESCRIPTION = [
+  'Крутая фотка',
+  'Если бы мои мысли можно было отразить на фотке, то это было бы так',
+  'Эмоции зашкаливают'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Джон', 'Стив', 'Илон', 'Кассандра', 'София', 'Эбигейл'];
+
+export {MAX_COUNT_PHOTOS, MAX_COUNT_COMMENTS, COUNT_LIKES, AVATAR_ID, DESCRIPTION, MESSAGES, NAMES};
