+++ js/utils.js
@@ -1,4 +1,3 @@
-// Функция взята с сайта https://up.htmlacademy.ru/univer-js1/1/tasks/7 и доработана.
 const getRandomNumber = (minimum, maximum) => {
   minimum = Math.ceil(Math.min(Math.abs(minimum), Math.abs(maximum)));
   maximum = Math.floor(Math.max(Math.abs(minimum), Math.abs(maximum)));
